1212last

+++ javawork/basicJava/src/workbook/pg55exam2/Test02.java
@@ -1,7 +1,7 @@
 package workbook.pg55exam2;
 public class Test02 {
 	public static void main(String[] args) {
-		Company company = new Company(Double.parseDouble(args[0]));
+		Company company = new Company(Integer.parseInt(args[0]));
 		company.print();
 	}
